apps/provider-integrations/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
urllib3==2.0.4
twilio==8.10.0
schedule==1.2.0
python-dateutil==2.8.2 